requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
colorama==0.4.6
python-dateutil==2.9.0
